Rename query/indidces to query/types
It didn't really contain a list of indices.
