+++ test/unit/query/types.js
@@ -1,12 +1,12 @@
 
-var indeces = require('../../../query/indeces');
+var types = require('../../../query/types');
 
 module.exports.tests = {};
 
 module.exports.tests.interface = function(test, common) {
   test('valid interface', function(t) {
-    t.true(Array.isArray(indeces), 'valid array');
-    t.equal(indeces.length, 11, 'valid array');
+    t.true(Array.isArray(types), 'valid array');
+    t.equal(types.length, 11, 'valid array');
     t.end();
   });
 };
@@ -14,7 +14,7 @@ module.exports.tests.interface = function(test, common) {
 module.exports.all = function (tape, common) {
 
   function test(name, testFunction) {
-    return tape('indeces ' + name, testFunction);
+    return tape('types ' + name, testFunction);
   }
 
   for( var testCase in module.exports.tests ){
